Decoupled tests of jsettlers.logic from the tests of jsettlers.common.
