+++ jsettlers.testutils/build.gradle
@@ -2,5 +2,7 @@ apply plugin: 'java'
 
 dependencies {
     compile 'junit:junit:4.12'
+    compile project(':jsettlers.common')
+    compile project(':jsettlers.logic')
     compile project(':jsettlers.main.swing')
 }
