--- jsettlers.testutils/test.gradle
@@ -1,1 +1,0 @@
-apply plugin: 'java'
